Fix connection leak in Hive table tests (#538)
This is another attempt to fix the connection leak in the Hive table tests. The current theory is that the cache in HiveCatalogs is not behaving correctly. Because it uses weak keys, garbage collection will almost always remove the cached catalog between SQL queries, and the next query will result in a new connection to the metastore. The garbage collected catalog is not closed in this case.

This fixes the leak in two ways:

1. Use a fixed duration after the last access, 10 minutes, to expire catalogs
2. Because catalogs are expired instead of garbage collected, use a removal listener to close
3. Add a catch-all finalizer to HiveCatalog to ensure that catalogs are closed when leaked, and a stack trace is logged to help fix the leak

(cherry picked from commit 5620f119f43d40cb657ed605ae6c647ee03e8a79)